added individual page css files and restructured into page folders with jsx and css. Added new Sessions page and linked it in Navbar and App
