+++ src/App.jsx
@@ -1,13 +1,14 @@
 import { useState } from 'react'
 import Navbar from "./Navbar.jsx";
 import './App.css'
-import Home from './pages/Home.jsx';
-import Player1 from './pages/Player1.jsx'
-import Player2 from './pages/Player2.jsx'
-import Recommendation from './pages/Recommendation.jsx';
-import SavedPlayers from './pages/SavedPlayers.jsx';
-import Summary from './pages/Summary.jsx'
-import Wishlist from './pages/Wishlist.jsx';
+import Home from './pages/Home/Home.jsx';
+import Player1 from './pages/Players/Player1.jsx'
+import Player2 from './pages/Players/Player2.jsx'
+import Recommendation from './pages/Recommendation/Recommendation.jsx';
+import SavedPlayers from './pages/SavedPlayers/SavedPlayers.jsx';
+import Sessions from './pages/Sessions/Sessions.jsx';
+import Summary from './pages/Summary/Summary.jsx'
+import Wishlist from './pages/Wishlist/Wishlist.jsx';
 import {Route, Routes} from 'react-router-dom';
 function App() {
  
@@ -23,6 +24,7 @@ function App() {
       <Route path='/player2' element ={<Player2 />} />
       <Route path='/recommendation' element ={<Recommendation />} />
       <Route path='/savedplayers' element ={<SavedPlayers />} />
+      <Route path='/sessions' element ={<Sessions />} />
       <Route path='/summary' element ={<Summary />} />
       <Route path='/wishlist' element ={<Wishlist />} />
     </Routes>
