+++ src/pages/Home/Home.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Home.css'
 
 const Home = () => {
   return (
