+++ src/pages/Players/Player1.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Player.css'
 
 const Player1 = () => {
   return (
