+++ src/pages/Players/Player2.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Player.css'
 
 const Player2 = () => {
   return (
