+++ src/pages/Recommendation/Recommendation.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Recommendation.css'
 
 const Recommendation = () => {
   return (
