+++ src/pages/Sessions/Sessions.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './Sessions.css'
+
+const Sessions = () => {
+  return (
+    <div>Sessions</div>
+  )
+}
+
+export default Sessions
