+++ src/pages/Summary/Summary.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Summary.css'
 
 const Summary = () => {
   return (
