+++ src/pages/Wishlist/Wishlist.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Wishlist.css'
 
 const Wishlist = () => {
   return (
